Refatora DTOs, validações e adiciona logs detalhados
Refatora `InvoiceDto` para usar enum `InvoiceStatus`, melhorando legibilidade e consistência. Ajusta validações no `InvoiceDtoValidator` para maior flexibilidade.

Adiciona logs detalhados com `Debug.WriteLine` em `InvoiceService`, `InvoiceRepository` e `PdvViewModel`, melhorando a rastreabilidade de erros e processos.

Atualiza `PdvView.xaml` com novos estilos para `ComboBox` e melhorias visuais. Ajusta lógica de estoque e finalização de vendas no `PdvViewModel`.

Garante consistência na serialização de valores no `ApplicationDbContext`. Melhora a experiência do usuário e a manutenção do sistema.

+++ VendaFlex/ViewModels/Dashboard/DashboardViewModel.cs
@@ -395,7 +395,7 @@ namespace VendaFlex.ViewModels.Dashboard
 
                 // Pagamentos Pendentes (Status Confirmed = 2 significa confirmado mas não totalmente pago)
                 var pendingInvoices = allInvoices
-                    .Where(i => i.Status == (int)InvoiceStatus.Confirmed)
+                    .Where(i => i.Status == InvoiceStatus.Confirmed)
                     .ToList();
 
                 var pendingAmount = pendingInvoices.Sum(i => i.Total - i.PaidAmount);
